+++ resources/views/layouts/app.blade.php
@@ -112,6 +112,7 @@
         <p>
             {{ config('app.name') }} &middot;
             <a href="{{ route('instance.rules') }}">{{ __('openbook.footer.rules') }}</a> &middot;
+            <a href="{{ route('instance.privacy') }}">{{ __('openbook.footer.privacy') }}</a> &middot;
             <a href="{{ config('openbook.homepage') }}">Openbook</a> v{{ config('openbook.version') }} &middot;
             {{ __('openbook.footer.license') }}
         </p>