+++ routes/web.php
@@ -20,6 +20,7 @@ use App\Http\Controllers\HashtagController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\InstanceRulesController;
 use App\Http\Controllers\LikeController;
+use App\Http\Controllers\PrivacyPolicyController;
 use App\Http\Controllers\MentionSuggestController;
 use App\Http\Controllers\NotificationController;
 use App\Http\Controllers\PostController;
@@ -32,6 +33,7 @@ use Illuminate\Support\Facades\Route;
 
 Route::get('/', [HomeController::class, 'index'])->name('home');
 Route::get('/regole', [InstanceRulesController::class, 'show'])->name('instance.rules');
+Route::get('/privacy', [PrivacyPolicyController::class, 'show'])->name('instance.privacy');
 
 Route::middleware('guest')->group(function () {
     Route::get('/registrati', [RegisterController::class, 'create'])->name('register');